refactor into its own file and fix formatting
